tests: re-organised the optional features test
