+++ extendr-api/tests/optional/mod.rs
@@ -1,2 +1,11 @@
 #[cfg(feature = "either")]
 mod either;
+
+#[cfg(feature = "ndarray")]
+mod ndarray;
+
+#[cfg(feature = "serde")]
+mod serde;
+
+#[cfg(feature = "non-api")]
+mod non_api;
