+++ extendr-api/tests/optional/non_api.rs
