+++ extendr-api/tests/optional/serde.rs
@@ -1,0 +1,2 @@
+mod deserializer;
+mod serializer;
